Just hydrate all data ahead of time
This obviates the need to distinguish between hydrated and unhydrated
types in most of the code and resolves a lot of typing headaches. It
creates a potential concern about memory usage because we're creating 4
Moment objects for every performance in the dataset. But, (1) it's
better to write the naive version first instead of prematurely
optimizing, and (2) Moment is deprecated anyway so, given that perf is
not an immediate empirical concern then I should wait until replacing
Moment to re-evaluate the potential perf issue.