Fix types and add auto-formatting (#52)
* wip

* format with prettier

* delete unused asset

* Almost first pass on types

- to make for a simpler starting place, I removed usage of React Context
  and am currently just managing state with a useReducer in the root
  component, and passing down state and dispatch functions through
  props. It's not a huge problem at the code's current scale. After I
  resolve the typing situation I could bring back useContext and
  simplify the prop passing.
- no more implicit any allowed
- added types for almost everything
- discovered flagsKey was missing an entry for the 'preview' flag so I
  think that was just never showing up. Thanks TS! :D
- discovered there was a sort of accidental divergence between the
  object shape for a performance being read from the data file vs. the
  shape expected by the render function. I haven't finished resolving
  this, so not only is the typechecker still complaining about it, but
  also I broke selection toggling in the current half-resolved state of
  this issue.

* Just hydrate all data ahead of time

This obviates the need to distinguish between hydrated and unhydrated
types in most of the code and resolves a lot of typing headaches. It
creates a potential concern about memory usage because we're creating 4
Moment objects for every performance in the dataset. But, (1) it's
better to write the naive version first instead of prematurely
optimizing, and (2) Moment is deprecated anyway so, given that perf is
not an immediate empirical concern then I should wait until replacing
Moment to re-evaluate the potential perf issue.